+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Add support for CORS in the HTTP Kafka Bridge
 * Pass HTTP Proxy configuration from operator to operands
 * Add Cruise Control support, KafkaRebalance resource and rebalance operator
+* Support dynamically changeable logging in the Entity Operator 
 
 ## 0.17.0